Reorganized models library and added StoryBuilderModel
